De-vectorize as much as possible.
For more objective testing, copy the contents of the original data into an
array that was manually heap allocated by me. This avoids any confusion
in terms of which STL allocator was used, etc. This is generally just a
better idea, since all my methods use a T* plus size for sorting anyway.
So we could even potentially load into a non-vector instead, and run
into no issues as far as testing goes.