Add variadic columns
Often it is useful to pass-through some data from the job generation
interface to the output CSV directly for a given run of QST. This commit
adds this feature with `--cols` and `--vals` as parameters. These
parameters are supplied as comma separated values to indicate which
additional columns to add in this CSV output, and what their values
should be.

For example:
```
$ ./QST ../data_super_mini/random/0.gz --cols a,b,c --vals 1,2,3 -o foo.csv

$ cat foo.csv
method,input,size,threshold,wall_nsecs,user_nsecs,system_nsecs,a,b,c
qsort_c,"../data_super_mini/random/0.gz",25,4,5301,0,0,1,2,3
```

`cols` are mapped in-order to `vals`. In this case, a->1, b->2, c->3.

This commit also tidies up the code a little bit, adding `const`
wherever possible and adding some additional parsers for the new CLI
flags.