Revamp of evaluator
A still in-progress, major revamp of the evaluator.

This commit introduces a separate evaluator interface for Google
benchmark results and my custom homebrewed QST results. Each type of
result is split into it's own set of loading, layout, and general files.

Still much to do, only a basic graph works for now.