Rework tests to use boost instead
This is a major overhaul to all the unit testing for the C part of this
codebase. Everything is now tested using Boosts parameterized tests
rather than the nightmare of macros that it was before. Most of the
logic is fairly self-explanatory and more maintainable. As a bonus, all
primitive types in C are now tested with all methods. The downside is
that this now requires a C++ compiler and boost to actually run the
tests, but for something this comprehensive, I think it is worth it.