Add C++ sorting function for testing
Add a small wrapper over std::sort to invoke the same C comparator and
present a C interface. This way we can build the C++ testing code in a
somewhat 'isolated' fashion and link to it and libstdc++ during the
final link phase.

This commit also cleans up quite a few warnings and _finally_ passes all
unit tests.